practicals/02_restful_api_development_with_flask/solution/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
waitress==3.0.0
Flask-SQLAlchemy==3.1.1